Fix MSVC warning C4244.
This warns about potential value truncation.  Instead of disabling
the warning, fix it, generally by adding explicit casts to make it
clear that a type change is occurring.

Reviewed-on: https://code-review.googlesource.com/1441

+++ re2/parse.cc
@@ -1147,7 +1147,7 @@ bool Regexp::ParseState::MaybeConcatString(int r, ParseFlags flags) {
   if (r >= 0) {
     re1->op_ = kRegexpLiteral;
     re1->rune_ = r;
-    re1->parse_flags_ = flags;
+    re1->parse_flags_ = static_cast<uint16>(flags);
     return true;
   }
